Add the list of environment variables method - code refactoring
