+++ examples/Domains/add_domain.js
@@ -1,4 +1,4 @@
-let scalingo = require('../dist/scalingo.js');
+let scalingo = require('../../dist/scalingo.js');
 
 scalingo.clientFromToken(process.env.SCALINGO_TOKEN).then((client)=>{
     return client.Domains.create("scalingo-js-tests", {name: "te.eplzzze"})
