+++ examples/Environment/list_environment.js
@@ -1,0 +1,7 @@
+let scalingo = require('../../dist/scalingo.js');
+
+scalingo.clientFromToken(process.env.SCALINGO_TOKEN).then((client)=>{
+  return client.Environment.for("scalingo-js-tests")
+}).then((domains) => {
+  console.log(domains)
+});
