+++ examples/User/get_user_infos.js
@@ -1,4 +1,4 @@
-let scalingo = require('../dist/scalingo.js')
+let scalingo = require('../../dist/scalingo.js');
 
 scalingo.clientFromToken(process.env.SCALINGO_TOKEN).then((client)=>{
   return client.Users.self()
