+++ src/index.js
@@ -4,6 +4,7 @@ import Metrics from './Metrics'
 import Tokens from './Tokens'
 import Users from './Users'
 import Domains from "./Domains"
+import Environment from "./Environment";
 
 import axios from 'axios'
 
@@ -54,6 +55,12 @@ class Client {
      * @type {Domains}
      */
     this.Domains = new Domains(this)
+  
+    /**
+     * Environment API
+     * @type {Environment}
+     */
+    this.Environment = new Environment(this)
   }
 
   /**
